Changed File attributes and constructor arguments
The File class no longer has the format or name arguments - format is
pointless (you know the format from the type) and name forces one to use
data in files, but as a library there may be use cases for handling raw
binary data (e.g. serialising and creating a polyglot, creating a nested
polyglot, etc.). The constructor now just takes data as an argument,
which is a string of bytes containing the data of the file.