Added CLI functionality to main program
There are several issues with the program in its current state (no
output messages to describe what's going on, some very confusing and
limited argument passing to functions, etc.) but it does work and can
produce JPEG/ZIP polyglots. Current plans are to expand to further
formats (PNG, GIF, GZIP, etc.) and then fix up the rest of the code.